requirements.txt
streamlit>=1.20.0
pandas
openai
openpyxl
psycopg2
python-dotenv
langchain-core
langchain-openai
langchain-community
sqlalchemy
tiktoken
altair<5
